requirements.txt
altair==4.1.0
appdirs==1.4.3
attrs==19.3.0
backcall==0.2.0
bcrypt==3.1.7
black==19.10b0
bleach==3.1.5
branca==0.4.1
brotlipy==0.7.0
cartotools==1.2.1
certifi==2020.6.20
cffi==1.14.0
cfgv==3.2.0
chardet==3.0.4
click==7.1.2
codecov==2.1.8
coverage==5.2.1
cryptography==2.9.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
entrypoints==0.3
filelock==3.0.12
flake8==3.8.3
future==0.18.2
identify==1.4.25
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1593446406207/work
iniconfig==1.0.1
ipykernel==5.3.0
ipyleaflet==0.13.1
ipython @ file:///tmp/build/80754af9/ipython_1593447377063/work
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort @ file:///tmp/build/80754af9/isort_1596472823065/work
jedi @ file:///tmp/build/80754af9/jedi_1592841866100/work
Jinja2==2.11.2
joblib @ file:///tmp/build/80754af9/joblib_1594236160679/work
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib @ file:///tmp/build/80754af9/matplotlib-base_1592846008246/work
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.1.0
mkl-random==1.1.1
mkl-service==2.3.0
more-itertools==8.4.0
MulticoreTSNE==0.1
mypy @ file:///tmp/build/80754af9/mypy_1593442617121/work
mypy-extensions==0.4.3
nbconvert==5.6.1
nbformat==5.0.7
nodeenv==1.4.0
notebook==6.0.3
numpy==1.18.5
olefile==0.46
opencv-python==4.3.0.36
OWSLib==0.19.2
packaging==20.4
pandas==1.0.5
pandocfilters==1.4.2
paramiko==2.7.1
parso==0.7.0
pathspec==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
pluggy==0.13.1
pre-commit==2.6.0
prometheus-client==0.8.0
prompt-toolkit==3.0.5
psutil==5.7.0
ptyprocess==0.6.0
py==1.9.0
pycodestyle==2.6.0
pycparser==2.20
pyepsg==0.4.0
pyflakes==2.2.0
Pygments==2.6.1
pykdtree==1.3.1
pyModeS==2.8
PyNaCl==1.4.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyproj==2.6.1.post1
pyrsistent==0.16.0
pyrtlsdr==0.2.92
pyshp==2.1.0
PySocks==1.7.1
pytest==6.0.1
pytest-cov==2.10.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.1
regex @ file:///tmp/build/80754af9/regex_1593435547652/work
requests @ file:///tmp/build/80754af9/requests_1592841827918/work
requests-pkcs12==1.7
scikit-learn @ file:///tmp/build/80754af9/scikit-learn_1592502866053/work
scipy @ file:///tmp/build/80754af9/scipy_1592930511789/work
Send2Trash==1.5.0
sip==4.19.13
six==1.15.0
styletrf @ file:///home/eisti/private/coding/projects/style-transfert
terminado==0.8.3
testpath==0.4.4
threadpoolctl @ file:///tmp/tmp9twdgx9k/threadpoolctl-2.1.0-py3-none-any.whl
toml @ file:///tmp/build/80754af9/toml_1592853716807/work
toolz==0.10.0
torch==1.5.1
torchvision==0.6.0a0+35d732a
tornado==6.0.4
tqdm==4.47.0
traitlets==4.3.3
traittypes==0.2.1
typed-ast==1.4.1
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1592847887441/work
urllib3==1.25.9
virtualenv==20.0.29
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.1.0
